requirements.txt
pygame
numpy
python-can
cantools
pysocket
pyserial
matplotlib
pomegranate
sphinx
sphinx-argparse
